+++ extensions/ExpansionPackage.lua
@@ -6624,359 +6624,6 @@ LuaNeifaTargetMod = sgs.CreateTargetModSkill {
 ExYuantanYuanshang:addSkill(LuaNeifa)
 SkillAnjiang:addSkill(LuaNeifaTargetMod)
 
-ExMouHuangzhong = sgs.General(extension, 'ExMouHuangzhong', 'shu', '4', true)
-
-LuaLiegong = sgs.CreateTriggerSkill {
-    name = 'LuaLiegong',
-    events = {sgs.TargetSpecified, sgs.DamageCaused, sgs.SlashProceed},
-    on_trigger = function(self, event, player, data, room)
-        if event == sgs.TargetSpecified then
-            local use = data:toCardUse()
-            local card = use.card
-            if use.to:length() > 1 then
-                return false
-            end
-            if card and card:isKindOf('Slash') then
-                local x = rinsan.getLiegongSuitNum(player)
-                if x > 0 then
-                    if room:askForSkillInvoke(player, self:objectName(), data) then
-                        room:broadcastSkillInvoke(self:objectName())
-                        room:setPlayerFlag(player, 'LuaLiegongInvoked')
-                        x = math.max(0, x - 1)
-                        if x > 0 then
-                            -- 参照源码【裸衣】亮出方式
-                            local card_ids = room:getNCards(x, false)
-                            for _, to in sgs.qlist(use.to) do
-                                room:setPlayerFlag(to, 'LuaLiegongTarget')
-                            end
-                            local move = sgs.CardsMoveStruct(card_ids, player, sgs.Player_PlaceTable,
-                                sgs.CardMoveReason(sgs.CardMoveReason_S_REASON_TURNOVER, player:objectName(),
-                                    self:objectName(), ''))
-                            room:moveCardsAtomic(move, true)
-                            room:getThread():delay()
-                            room:getThread():delay()
-                            local dummy = sgs.Sanguosha:cloneCard('slash', sgs.Card_NoSuit, 0)
-                            local damage_count = 0
-                            for _, card_id in sgs.qlist(card_ids) do
-                                local cd = sgs.Sanguosha:getCard(card_id)
-                                local suit_string = rinsan.firstToUpper(cd:getSuitString())
-                                if player:getMark('@LuaLiegong' .. suit_string) > 0 then
-                                    damage_count = damage_count + 1
-                                end
-                                dummy:addSubcard(cd)
-                            end
-                            card:setTag('LuaLiegongExtraDamage', sgs.QVariant(damage_count))
-                            room:throwCard(dummy, sgs.CardMoveReason(sgs.CardMoveReason_S_REASON_NATURAL_ENTER,
-                                player:objectName(), self:objectName(), ''), nil)
-                        end
-                    end
-                end
-            end
-        elseif event == sgs.DamageCaused then
-            local damage = data:toDamage()
-            if damage.transfer or damage.chain then
-                return false
-            end
-            local card = damage.card
-            if card and card:isKindOf('Slash') then
-                local x = card:getTag('LuaLiegongExtraDamage'):toInt()
-                if x > 0 then
-                    damage.damage = damage.damage + x
-                    room:sendCompulsoryTriggerLog(player, self:objectName())
-                    data:setValue(damage)
-                end
-                card:removeTag('LuaLiegongExtraDamage')
-            end
-        else
-            local effect = data:toSlashEffect()
-            if not effect.to:hasFlag('LuaLiegongTarget') then
-                return false
-            end
-            local all_suits = {'heart', 'diamond', 'club', 'spade'}
-            local suits = {}
-            for _, suit in ipairs(all_suits) do
-                if effect.from:getMark('@LuaLiegong' .. rinsan.firstToUpper(suit)) > 0 then
-                    table.insert(suits, suit)
-                end
-            end
-            if #suits > 0 then
-                room:setPlayerCardLimitation(effect.to, 'use, response', 'Jink|' .. table.concat(suits, ','), false)
-                local source = room:findPlayerBySkillName(self:objectName())
-                local prompt = string.format('@LuaLiegong-jink:%s:%s:%s', effect.from:objectName(), source:objectName(),
-                    self:objectName())
-                local jink = room:askForCard(effect.to, 'jink', prompt, data, sgs.Card_MethodUse, source)
-                if jink then
-                    local invalid_jink = table.contains(suits, jink:getSuitString())
-                    if invalid_jink then
-                        rinsan.sendLogMessage(room, '#LuaLiegongInvalidJink', {
-                            ['from'] = effect.to,
-                            ['card_str'] = jink:toString()
-                        })
-                        room:slashResult(effect, nil)
-                    else
-                        room:slashResult(effect, jink)
-                    end
-                else
-                    room:slashResult(effect, nil)
-                end
-                room:removePlayerCardLimitation(effect.to, 'use, response', 'Jink|' .. table.concat(suits, ','))
-                return true
-            end
-        end
-        return false
-    end,
-    can_trigger = function(self, target)
-        return target
-    end
-}
-
-LuaLiegongMark = sgs.CreateTriggerSkill {
-    name = 'LuaLiegongMark',
-    events = {sgs.CardUsed, sgs.TargetConfirmed, sgs.CardFinished, sgs.CardResponded},
-    global = true,
-    on_trigger = function(self, event, player, data, room)
-        if event == sgs.CardFinished then
-            if player:hasFlag('LuaLiegongInvoked') then
-                room:setPlayerFlag(player, '-LuaLiegongInvoked')
-                room:setPlayerMark(player, '@LuaLiegongHeart', 0)
-                room:setPlayerMark(player, '@LuaLiegongClub', 0)
-                room:setPlayerMark(player, '@LuaLiegongSpade', 0)
-                room:setPlayerMark(player, '@LuaLiegongDiamond', 0)
-            end
-        elseif event == sgs.TargetConfirmed then
-            local use = data:toCardUse()
-            local card = use.card
-            if rinsan.cardCanBeRecorded(card) then
-                if use.to:contains(player) and player:hasSkill('LuaLiegong') then
-                    room:setPlayerMark(player, rinsan.getLiegongSuitMarkName(card), 1)
-                end
-            end
-        else
-            local card
-            if event == sgs.CardUsed then
-                card = data:toCardUse().card
-            else
-                -- 用于区分无目标响应使用卡牌，例如【闪】
-                local resp = data:toCardResponse()
-                if not resp.m_isUse then
-                    return false
-                end
-                card = resp.m_card
-            end
-            if rinsan.cardCanBeRecorded(card) then
-                if player:hasSkill('LuaLiegong') then
-                    room:setPlayerMark(player, rinsan.getLiegongSuitMarkName(card), 1)
-                end
-            end
-        end
-    end,
-    can_trigger = function(self, target)
-        return target
-    end
-}
-
-LuaLiegongAttackMod = sgs.CreateTargetModSkill {
-    name = 'LuaLiegongAttackMod',
-    pattern = 'Slash',
-    distance_limit_func = function(self, from, card, to)
-        if from:hasSkill('LuaLiegong') then
-            return math.max(card:getNumber() - from:getAttackRange(), 0)
-        end
-    end
-}
-
-ExMouHuangzhong:addSkill(LuaLiegong)
-SkillAnjiang:addSkill(LuaLiegongAttackMod)
-SkillAnjiang:addSkill(LuaLiegongMark)
-
-ExMouGanning = sgs.General(extension, 'ExMouGanning', 'wu', '4', true, true)
-
-LuaQixi = sgs.CreateOneCardViewAsSkill {
-    name = 'LuaQixi',
-    filter_pattern = '.|black',
-    view_as = function(self, card)
-        local acard = sgs.Sanguosha:cloneCard('dismantlement', card:getSuit(), card:getNumber())
-        acard:addSubcard(card)
-        acard:setSkillName(self:objectName())
-        return acard
-    end
-}
-
-LuaQixiTrigger = sgs.CreateTriggerSkill {
-    name = 'LuaQixiTrigger',
-    events = {sgs.TrickEffect},
-    global = true,
-    on_trigger = function(self, event, player, data, room)
-        local effect = data:toCardEffect()
-        local card = effect.card
-        -- 替代的【过河拆桥】判断
-        if (card:isKindOf('Dismantlement') or card:objectName() == 'dismantlement') and (not card:isVirtualCard()) then
-            if not effect.from:hasSkill('LuaQixi') then
-                return false
-            end
-            if (room:askForSkillInvoke(effect.from, 'LuaQixi', data)) then
-                local dummy = sgs.Sanguosha:cloneCard('slash', sgs.Card_NoSuit, -1)
-                for _, cd in sgs.qlist(effect.to:getCards('hej')) do
-                    if (rinsan.canDiscardCard(effect.from, effect.to, cd:getId())) then
-                        dummy:addSubcard(cd)
-                    end
-                end
-                if dummy:subcardsLength() > 0 then
-                    room:throwCard(dummy, effect.to, effect.from)
-                end
-                room:broadcastSkillInvoke('LuaQixi')
-                return true
-            end
-        end
-        return false
-    end,
-    can_trigger = function(self, target)
-        return target
-    end
-}
-
-LuaFenweiCard = sgs.CreateSkillCard {
-    name = 'LuaFenweiCard',
-    target_fixed = false,
-    filter = function(self, targets, to_select, player)
-        return to_select:hasFlag('trickcardtarget')
-    end,
-    on_use = function(self, room, source, targets)
-        source:loseMark('@fenwei')
-        for _, p in ipairs(targets) do
-            room:setPlayerFlag(p, 'luafenwei')
-        end
-    end
-}
-
-LuaFenweiVS = sgs.CreateViewAsSkill {
-    name = 'LuaFenwei',
-    n = 0,
-    view_as = function(self, cards)
-        return LuaFenweiCard:clone()
-    end,
-    enabled_at_play = function(self, player)
-        return false
-    end,
-    enabled_at_response = function(self, player, pattern)
-        return player:getMark('@fenwei') > 0 and pattern == '@@LuaFenwei'
-    end
-}
-
-LuaFenwei = sgs.CreateTriggerSkill {
-    name = 'LuaFenwei',
-    frequency = sgs.Skill_Limited,
-    limit_mark = '@fenwei',
-    events = {sgs.CardUsed},
-    view_as_skill = LuaFenweiVS,
-    on_trigger = function(self, event, player, data, room)
-        local use = data:toCardUse()
-        if not use.card:isKindOf('TrickCard') then
-            return false
-        end
-        if use.to:length() < 2 then
-            return false
-        end
-        local splayer = room:findPlayerBySkillName(self:objectName())
-        if splayer then
-            for _, dest in sgs.qlist(use.to) do
-                room:setPlayerFlag(dest, 'trickcardtarget')
-            end
-            if room:askForUseCard(splayer, '@@LuaFenwei', '@LuaFenwei') then
-                local targetCount = 0
-                local nullified_list = use.nullified_list
-                for _, dest in sgs.qlist(use.to) do
-                    if dest:hasFlag('luafenwei') then
-                        table.insert(nullified_list, dest:objectName())
-                        room:setPlayerFlag(dest, '-luafenwei')
-                        targetCount = targetCount + 1
-                    end
-                end
-                targetCount = math.min(4, targetCount)
-                use.nullified_list = nullified_list
-                data:setValue(use)
-                room:broadcastSkillInvoke(self:objectName())
-                room:setEmotion(splayer, 'skill/ganning_fenwei')
-                local checker = function(card)
-                    return card:isKindOf('Dismantlement')
-                end
-                while targetCount > 0 do
-                    local dismantlement = rinsan.obtainCardFromPile(checker, room:getDrawPile())
-                    if dismantlement then
-                        splayer:obtainCard(dismantlement)
-                    else
-                        break
-                    end
-                    targetCount = targetCount - 1
-                end
-            end
-            for _, target in sgs.qlist(room:getAllPlayers()) do
-                room:setPlayerFlag(target, '-trickcardtarget')
-            end
-        end
-        return false
-    end,
-    can_trigger = function(self, target)
-        return true
-    end
-}
-
--- 替换原版【过河拆桥】
-LuaDismantlement = sgs.CreateTrickCard {
-    name = 'dismantlement',
-    subtype = 'single_target_trick',
-    filter = function(self, targets, to_select, player)
-        local total_num = sgs.Sanguosha:correctCardTarget(sgs.TargetModSkill_ExtraTarget, player, self)
-        return
-            targets:length() < total_num and to_select ~= player:objectName() and to_select:getCards('hej'):length() > 0
-    end,
-    on_effect = function(self, effect)
-        local source = effect.from
-        local target = effect.to
-        if source:isDead() then
-            return
-        end
-        local room = effect.to:getRoom()
-        if (not rinsan.canDiscard(source, target, 'hej')) then
-            return
-        end
-        local card_id = room:askForCardChosen(source, target, 'hej', self:objectName(), false, sgs.Card_MethodDiscard)
-        local place = target
-        if room:getCardPlace(card_id) == sgs.Player_PlaceDelayedTrick then
-            place = nil
-        end
-        room:throwCard(card_id, place, source)
-    end
-}
-
-LuaQicaiHotfix = sgs.CreateTriggerSkill {
-    name = 'LuaQicaiHotfix',
-    frequency = sgs.Skill_Compulsory,
-    priority = 10000,
-    global = true,
-    events = {sgs.CardEffected},
-    on_trigger = function(self, event, player, data, room)
-        local effect = data:toCardEffect()
-        if effect.to:hasSkill('qicai') and effect.card and effect.card:isKindOf('Dismantlement') then
-            local dismantlement = LuaDismantlement:clone()
-            dismantlement:setSuit(effect.card:getSuit())
-            dismantlement:setNumber(effect.card:getNumber())
-            dismantlement:setId(effect.card:getId())
-            effect.card = dismantlement
-            data:setValue(effect)
-        end
-        return false
-    end,
-    can_trigger = function(self, target)
-        return target
-    end
-}
-
-ExMouGanning:addSkill(LuaQixi)
-ExMouGanning:addSkill(LuaFenwei)
-SkillAnjiang:addSkill(LuaQixiTrigger)
-SkillAnjiang:addSkill(LuaQicaiHotfix)
-
 JieJiaxu = sgs.General(extension, 'JieJiaxu', 'qun', '3', true)
 
 LuaWansha = sgs.CreateTriggerSkill {
@@ -8035,130 +7682,6 @@ SkillAnjiang:addSkill(LuaFuhaiClear)
 SkillAnjiang:addSkill(LuaFuhaiDeath)
 ExShenSunce:addSkill(LuaPinghe)
 SkillAnjiang:addSkill(LuaPingheMaxCards)
-
-ExMouMachao = sgs.General(extension, 'ExMouMachao', 'shu', '4', true)
-
--- 判断谋弈是否成功
-local function checkLuaMouTiejiMouyi(sourceChoice, targetChoice)
-    return string.sub(sourceChoice, -1) ~= string.sub(targetChoice, -1)
-end
-
-LuaMouTieji = sgs.CreateTriggerSkill {
-    name = 'LuaMouTieji',
-    frequency = sgs.Skill_NotFrequent,
-    events = {sgs.TargetSpecifying, sgs.TargetSpecified},
-    on_trigger = function(self, event, player, data, room)
-        local use = data:toCardUse()
-        if (not use.card) or (not use.card:isKindOf('Slash')) then
-            return false
-        end
-        if event == sgs.TargetSpecifying then
-            local index = 1
-            local indexes = sgs.IntList()
-            for _, p in sgs.qlist(use.to) do
-                if not player:isAlive() then
-                    break
-                end
-                local data2 = sgs.QVariant()
-                data2:setValue(p)
-                if room:askForSkillInvoke(player, self:objectName(), data2) then
-                    indexes:append(index)
-                    room:addPlayerMark(p, 'LuaMouTieji')
-                    room:addPlayerMark(p, '@skill_invalidity')
-                    room:doAnimate(1, player:objectName(), p:objectName())
-                    room:broadcastSkillInvoke(self:objectName(), 1)
-                    local sourceChoice = room:askForChoice(player, self:objectName(),
-                        'LuaMouTiejiAttack1+LuaMouTiejiAttack2', data2)
-                    local targetChoice = room:askForChoice(p, self:objectName(),
-                        'LuaMouTiejiDefense1+LuaMouTiejiDefense2')
-                    rinsan.sendLogMessage(room, '#choose', {
-                        ['from'] = player,
-                        ['arg'] = sourceChoice
-                    })
-                    rinsan.sendLogMessage(room, '#choose', {
-                        ['from'] = p,
-                        ['arg'] = targetChoice
-                    })
-                    local success = checkLuaMouTiejiMouyi(sourceChoice, targetChoice)
-                    local type = success and '#LuaMouTiejiSuccess' or '#LuaMouTiejiFailure'
-                    rinsan.sendLogMessage(room, type, {
-                        ['from'] = player,
-                        ['arg'] = 'LuaMouTiejiMouyi'
-                    })
-                    if success then
-                        if sourceChoice == 'LuaMouTiejiAttack1' then
-                            room:broadcastSkillInvoke(self:objectName(), 2)
-                            if p:isNude() then
-                                return false
-                            end
-                            local card_id = room:askForCardChosen(player, p, 'he', self:objectName(), false,
-                                sgs.Card_MethodNone)
-
-                            local reason = sgs.CardMoveReason(sgs.CardMoveReason_S_REASON_EXTRACTION,
-                                player:objectName())
-                            room:obtainCard(player, sgs.Sanguosha:getCard(card_id), reason, false)
-                        else
-                            room:broadcastSkillInvoke(self:objectName(), 3)
-                            player:drawCards(2, self:objectName())
-                        end
-                    else
-                        room:broadcastSkillInvoke(self:objectName(), 4)
-                    end
-                    if p:isAlive() then
-                        rinsan.sendLogMessage(room, '#NoJink', {
-                            ['from'] = p
-                        })
-                    end
-                end
-                index = index + 1
-            end
-            if indexes:length() == 0 then
-                return false
-            end
-            room:setPlayerFlag(player, 'LuaMouTiejiInvoked')
-            local indexes_data = sgs.QVariant()
-            indexes_data:setValue(indexes)
-            player:setTag('LuaMouTiejiTargets', indexes_data)
-            return false
-        end
-        if not player:hasFlag('LuaMouTiejiInvoked') then
-            return false
-        end
-        local jink_table = sgs.QList2Table(player:getTag('Jink_' .. use.card:toString()):toIntList())
-        local indexes = player:getTag('LuaMouTiejiTargets'):toIntList()
-        for _, index in sgs.qlist(indexes) do
-            jink_table[index] = 0
-        end
-        local jink_data = sgs.QVariant()
-        jink_data:setValue(Table2IntList(jink_table))
-        player:setTag('Jink_' .. use.card:toString(), jink_data)
-        player:removeTag('LuaMouTiejiTargets')
-        room:setPlayerFlag(player, '-LuaMouTiejiInvoked')
-        return false
-    end
-}
-
-LuaMouTiejiClear = sgs.CreateTriggerSkill {
-    name = 'LuaMouTiejiClear',
-    events = {sgs.EventPhaseChanging},
-    global = true,
-    on_trigger = function(self, event, player, data, room)
-        if data:toPhaseChange().to == sgs.Player_NotActive then
-            for _, p in sgs.qlist(room:getAlivePlayers()) do
-                local x = p:getMark('LuaMouTieji')
-                room:removePlayerMark(p, 'LuaMouTieji', x)
-                room:removePlayerMark(p, '@skill_invalidity', x)
-            end
-        end
-    end,
-    can_trigger = function(self, target)
-        return true
-    end
-}
-
-ExMouMachao:addSkill('mashu')
-ExMouMachao:addSkill(LuaMouTieji)
-SkillAnjiang:addSkill(LuaMouTiejiClear)
 
 -- 初始随机魏国/吴国
 local wenyang_kingdoms = {'wei', 'wu'}
