+++ lang/zh_CN/Audio/ExpansionPackageLines.lua
@@ -283,16 +283,6 @@ return {
     ['$LuaNeifa1'] = '同室内伐，贻笑外人',
     ['$LuaNeifa2'] = '自相恩残，相煎何急',
     ['~ExYuantanYuanshang'] = '兄弟难齐心，该有此果……',
-    -- 谋黄忠
-    ['$LuaLiegong1'] = '矢贯坚石，劲冠三军！',
-    ['$LuaLiegong2'] = '吾虽年迈，箭矢犹锋！',
-    ['~ExMouHuangzhong'] = '弦断弓藏，将老孤亡…',
-    -- 谋甘宁
-    ['$LuaQixi1'] = '击敌不备，奇袭拔寨！',
-    ['$LuaQixi2'] = '轻羽透重铠，奇袭溃坚城！',
-    ['$LuaFenwei1'] = '舍身护主，扬吴将之风！',
-    ['$LuaFenwei2'] = '袭军挫阵，奋江东之威！',
-    ['~ExMouGanning'] = '蛮将休得猖狂！呃啊！',
     -- 界贾诩
     ['$LuaLuanwu1'] = '一切…都在我的掌控中！',
     ['$LuaLuanwu2'] = '这乱世还不够乱！',
@@ -335,12 +325,6 @@ return {
     ['$LuaPinghe1'] = '不过胆小鼠辈，吾等有何惧哉！',
     ['$LuaPinghe2'] = '只可得胜而返，岂能败战而归！',
     ['~ExShenSunce'] = '无耻小人！胆敢暗算于我……',
-    -- 谋马超
-    ['$LuaMouTieji1'] = '厉兵秣马，只待今日！', -- 发动技能
-    ['$LuaMouTieji2'] = '敌军防备空虚，出击直取敌营！', -- 直取敌营
-    ['$LuaMouTieji3'] = '敌军早有防备，先行扰阵疲敌！', -- 扰阵疲敌
-    ['$LuaMouTieji4'] = '全军速撤回营，以期再觅良机！', -- 谋弈失败
-    ['~ExMouMachao'] = '父亲！不能为汝报仇雪恨矣……',
     -- 文鸯
     ['$LuaQuedi1'] = '力摧敌阵，如视天光破云！',
     ['$LuaQuedi2'] = '让尔等有命追，无命回！',
