+++ lang/zh_CN/Audio/ShieldPackageLines.lua
@@ -1,53 +1,3 @@
--- lines for Shield Package
+-- lines for ShieldPackage
 
-return {
-    -- 谋吕蒙
-    ['$LuaMouKeji1'] = '事事克己，步步虚心！',
-    ['$LuaMouKeji2'] = '勤学潜习，始觉自新！',
-    ['$LuaMouDujiang1'] = '大军浮江，昼夜驰上！',
-    ['$LuaMouDujiang2'] = '白衣摇橹，昼夜兼行！',
-    ['$LuaMouDuojing1'] = '快舟轻甲，速袭其后！',
-    ['$LuaMouDuojing2'] = '复取荆州，尽在掌握！',
-    ['~ExMouLvmeng'] = '义封胆略过人，主公可任之……',
-    -- 谋于禁
-    ['$LuaMouXiayuan1'] = '速置粮草，驰援天柱山！',
-    ['$LuaMouXiayuan2'] = '援军既至，定攻克此地！',
-    ['$LuaMouJieyue1'] = '尔等小儿，徒费兵力',
-    ['$LuaMouJieyue2'] = '雕虫小技，静待则已！',
-    ['~ExMouYujin'] = '禁……愧于丞相……',
-    -- 谋华雄
-    ['$LuaMouYaowu1'] = '俞涉小儿，岂是我的对手！',
-    ['$LuaMouYaowu2'] = '上将潘凤？哼！还不是死在我刀下！',
-    ['$LuaMouYangwei1'] = '哈哈哈哈！现在谁不知我华雄？',
-    ['$LuaMouYangwei2'] = '定要关外诸侯，知我威名！',
-    ['~ExMouHuaxiong'] = '小小马弓手，竟然……啊……',
-    -- 谋孙尚香
-    ['$LuaMouLiangzhu1'] = '助君得胜战，跃马提缨枪！',
-    ['$LuaMouLiangzhu2'] = '平贼成君业，何惜上沙场！',
-    ['$LuaMouJieyin1'] = '君若不负吾心，妾自随君千里！',
-    ['$LuaMouJieyin2'] = '夫妻之情既断，何必再问归期！',
-    ['$LuaMouXiaoji1'] = '吾之所通，何止十八般兵刃！',
-    ['$LuaMouXiaoji2'] = '既如此，就让尔等见识一番！',
-    ['~ExMouSunshangxiang'] = '此去一别，竟无再见之日……',
-    -- 谋黄盖
-    ['$LuaMouKurou1'] = '既不能破，不如依张子布之言，投降便罢！',
-    ['$LuaMouKurou2'] = '周瑜小儿！破曹不得，便欺吾三世老臣乎？',
-    ['$LuaMouZhaxiang1'] = '江东六郡之卒，怎敌丞相百万雄师！',
-    ['$LuaMouZhaxiang2'] = '闻丞相虚心纳士，盖愿率众归降！',
-    ['~ExMouHuanggai'] = '哈哈哈哈，公瑾计成，老夫死也无憾了……',
-    -- 谋曹仁
-    ['$LuaMouJushou1'] = '白马沉河共歃誓，怒涛没城亦不悔！',
-    ['$LuaMouJushou2'] = '山水速疾来去易，襄樊镇固永难开！',
-    ['$LuaMouJushou3'] = '汉水溢流断归路，守城之志穷且坚！',
-    ['$LuaMouJiewei1'] = '同袍之谊，断不可弃之！',
-    ['$LuaMouJiewei2'] = '贼虽势盛，若吾出马，亦可解之',
-    ['~ExMouCaoren'] = '吾身可殉，然襄樊之地万不可落于吴蜀之手……',
-    -- 谋徐晃
-    ['$LuaMouDuanliang1'] = '常读兵法，终有良策也！', -- 发动技能
-    ['$LuaMouDuanliang2'] = '烧敌粮草，救主于危急！', -- 围城断粮
-    ['$LuaMouDuanliang3'] = '敌陷混乱之机，我军可长驱直入！', -- 擂鼓进军
-    ['$LuaMouDuanliang4'] = '敌既识破吾计，则断不可行矣！', -- 谋弈失败
-    ['$LuaMouShipo1'] = '已向尔等陈明利害，奉劝尔等早日归降！',
-    ['$LuaMouShipo2'] = '此时归降或可封赏，即至城破必斩无赦！',
-    ['~ExMouXuhuang'] = '为主效劳，何畏生死……'
-}
+return {}
