+++ lang/zh_CN/Audio/StrategicAttackAbilityPackageLines.lua
@@ -1,0 +1,18 @@
+-- lines for StrategicAttackAbilityPackage
+
+return {
+    -- 谋华雄
+    ['$LuaMouYaowu1'] = '俞涉小儿，岂是我的对手！',
+    ['$LuaMouYaowu2'] = '上将潘凤？哼！还不是死在我刀下！',
+    ['$LuaMouYangwei1'] = '哈哈哈哈！现在谁不知我华雄？',
+    ['$LuaMouYangwei2'] = '定要关外诸侯，知我威名！',
+    ['~ExMouHuaxiong'] = '小小马弓手，竟然……啊……',
+    -- 谋孙尚香
+    ['$LuaMouLiangzhu1'] = '助君得胜战，跃马提缨枪！',
+    ['$LuaMouLiangzhu2'] = '平贼成君业，何惜上沙场！',
+    ['$LuaMouJieyin1'] = '君若不负吾心，妾自随君千里！',
+    ['$LuaMouJieyin2'] = '夫妻之情既断，何必再问归期！',
+    ['$LuaMouXiaoji1'] = '吾之所通，何止十八般兵刃！',
+    ['$LuaMouXiaoji2'] = '既如此，就让尔等见识一番！',
+    ['~ExMouSunshangxiang'] = '此去一别，竟无再见之日……'
+}
