+++ lang/zh_CN/Audio/StrategicAttackKnowledgePackageLines.lua
@@ -1,0 +1,17 @@
+-- lines for StrategicAttackKnowledgePackage
+return {
+    -- 谋徐晃
+    ['$LuaMouDuanliang1'] = '常读兵法，终有良策也！', -- 发动技能
+    ['$LuaMouDuanliang2'] = '烧敌粮草，救主于危急！', -- 围城断粮
+    ['$LuaMouDuanliang3'] = '敌陷混乱之机，我军可长驱直入！', -- 擂鼓进军
+    ['$LuaMouDuanliang4'] = '敌既识破吾计，则断不可行矣！', -- 谋弈失败
+    ['$LuaMouShipo1'] = '已向尔等陈明利害，奉劝尔等早日归降！',
+    ['$LuaMouShipo2'] = '此时归降或可封赏，即至城破必斩无赦！',
+    ['~ExMouXuhuang'] = '为主效劳，何畏生死……',
+    -- 谋马超
+    ['$LuaMouTieji1'] = '厉兵秣马，只待今日！', -- 发动技能
+    ['$LuaMouTieji2'] = '敌军防备空虚，出击直取敌营！', -- 直取敌营
+    ['$LuaMouTieji3'] = '敌军早有防备，先行扰阵疲敌！', -- 扰阵疲敌
+    ['$LuaMouTieji4'] = '全军速撤回营，以期再觅良机！', -- 谋弈失败
+    ['~ExMouMachao'] = '父亲！不能为汝报仇雪恨矣……'
+}
