+++ lang/zh_CN/Audio/StrategicAttackShacklePackageLines.lua
@@ -1,0 +1,41 @@
+-- lines for StrategicAttackShacklePackage
+
+return {
+    -- 谋于禁
+    ['$LuaMouXiayuan1'] = '速置粮草，驰援天柱山！',
+    ['$LuaMouXiayuan2'] = '援军既至，定攻克此地！',
+    ['$LuaMouJieyue1'] = '尔等小儿，徒费兵力',
+    ['$LuaMouJieyue2'] = '雕虫小技，静待则已！',
+    ['~ExMouYujin'] = '禁……愧于丞相……',
+    -- 谋吕蒙
+    ['$LuaMouKeji1'] = '事事克己，步步虚心！',
+    ['$LuaMouKeji2'] = '勤学潜习，始觉自新！',
+    ['$LuaMouDujiang1'] = '大军浮江，昼夜驰上！',
+    ['$LuaMouDujiang2'] = '白衣摇橹，昼夜兼行！',
+    ['$LuaMouDuojing1'] = '快舟轻甲，速袭其后！',
+    ['$LuaMouDuojing2'] = '复取荆州，尽在掌握！',
+    ['~ExMouLvmeng'] = '义封胆略过人，主公可任之……',
+    -- 谋曹仁
+    ['$LuaMouJushou1'] = '白马沉河共歃誓，怒涛没城亦不悔！',
+    ['$LuaMouJushou2'] = '山水速疾来去易，襄樊镇固永难开！',
+    ['$LuaMouJushou3'] = '汉水溢流断归路，守城之志穷且坚！',
+    ['$LuaMouJiewei1'] = '同袍之谊，断不可弃之！',
+    ['$LuaMouJiewei2'] = '贼虽势盛，若吾出马，亦可解之',
+    ['~ExMouCaoren'] = '吾身可殉，然襄樊之地万不可落于吴蜀之手……',
+    -- 谋甘宁
+    ['$LuaQixi1'] = '击敌不备，奇袭拔寨！',
+    ['$LuaQixi2'] = '轻羽透重铠，奇袭溃坚城！',
+    ['$LuaFenwei1'] = '舍身护主，扬吴将之风！',
+    ['$LuaFenwei2'] = '袭军挫阵，奋江东之威！',
+    ['~ExMouGanning'] = '蛮将休得猖狂！呃啊！',
+    -- 谋黄忠
+    ['$LuaLiegong1'] = '矢贯坚石，劲冠三军！',
+    ['$LuaLiegong2'] = '吾虽年迈，箭矢犹锋！',
+    ['~ExMouHuangzhong'] = '弦断弓藏，将老孤亡…',
+    -- 谋黄盖
+    ['$LuaMouKurou1'] = '既不能破，不如依张子布之言，投降便罢！',
+    ['$LuaMouKurou2'] = '周瑜小儿！破曹不得，便欺吾三世老臣乎？',
+    ['$LuaMouZhaxiang1'] = '江东六郡之卒，怎敌丞相百万雄师！',
+    ['$LuaMouZhaxiang2'] = '闻丞相虚心纳士，盖愿率众归降！',
+    ['~ExMouHuanggai'] = '哈哈哈哈，公瑾计成，老夫死也无憾了……'
+}
