+++ lang/zh_CN/Package/ExpansionPackage.lua
@@ -568,25 +568,6 @@ return {
     ['LuaNeifa-invoke'] = '你可以选择一名角色，使其成为或不再是 %src 的目标',
     ['#LuaNeifaAppend'] = '%to 成为了 %card 的目标',
     ['#LuaNeifaRemove'] = '%to 不再是 %card 的目标',
-    ['ExMouHuangzhong'] = '谋黄忠',
-    ['&ExMouHuangzhong'] = '谋黄忠',
-    ['#ExMouHuangzhong'] = '没金铩羽',
-    ['LuaLiegong'] = '烈弓',
-    [':LuaLiegong'] = '①你使用【杀】时可以选择距离不大于此【杀】点数的角色为目标；\
-    ②当你使用牌时，或成为其他角色使用牌的目标后，若此牌的花色未被记录，记录此花色；\
-    ③当你使用【杀】指定一名角色为唯一目标后，你可以展示牌堆顶的X张牌（X为你已记录的花色数-1，且至少为0），其中每有一张已记录的花色的牌，此【杀】的伤害便+1，且其不能使用已记录的花色的牌响应此【杀】，然后此【杀】结算结束后，清除所有已记录的花色',
-    ['@LuaLiegong-jink'] = '%src 使用了【杀】，请使用一张未被其“烈弓”记录过花色的【闪】',
-    ['#LuaLiegongInvalidJink'] = '由于已被记录花色，%from 使用的 %card 无效',
-    ['ExMouGanning'] = '谋甘宁',
-    ['&ExMouGanning'] = '谋甘宁',
-    ['#ExMouGanning'] = '兴王定霸',
-    ['LuaQixi'] = '奇袭',
-    [':LuaQixi'] = '你可以将黑色牌当作【过河拆桥】使用，你使用的非转化非虚拟【过河拆桥】可以弃置目标区域内所有的牌',
-    ['LuaFenwei'] = '奋威',
-    [':LuaFenwei'] = '限定技，当一张锦囊牌指定至少两个目标后，你可以令此牌对其中任意名目标角色无效，然后你从牌堆中获得X张【过河拆桥】（X为你以此法选择的角色数，且至多为4）',
-    ['luafenwei'] = '奋威',
-    ['@LuaFenwei'] = '你可以发动“奋威”',
-    ['~LuaFenwei'] = '选择若干名角色→点击确定',
     ['JieJiaxu'] = '界贾诩',
     ['&JieJiaxu'] = '界贾诩',
     ['#JieJiaxu'] = '冷酷的毒士',
@@ -662,24 +643,6 @@ return {
     [':LuaPinghe'] = '锁定技，你的手牌上限为你已损失的体力值。当你受到其他角色造成的伤害时，若你有手牌且体力上限大于1，你防止此伤害，然后你减1点体力上限并将一张手牌交给一名其他角色，然后若你拥有“英霸”，你令伤害来源获得1个“平定”标记',
     ['LuaPingheGive'] = '“冯河”被触发，你须交给其他角色一张手牌',
     ['~LuaPinghe'] = '选择一张手牌→选择一名其他角色→点击“确定”',
-    ['ExMouMachao'] = '谋马超',
-    ['&ExMouMachao'] = '谋马超',
-    ['#ExMouMachao'] = '阻戎负勇',
-    ['LuaMouTiejiMouyi'] = '铁骑谋弈',
-    ['LuaMouTieji'] = '铁骑',
-    [':LuaMouTieji'] = '当你使用【杀】指定一名角色为目标时，你可以令其所有非锁定技失效直到回合结束，且其不能使用【闪】响应此【杀】，使用然后你与该角色进行一次“谋弈”：\
-    1. 直取敌营，你获得其一张牌；\
-    2. 扰阵疲敌，你摸两张牌',
-    ['LuaMouTiejiAttack1'] = '直取敌营',
-    [':LuaMouTiejiAttack1'] = '若成功，你获得其一张牌',
-    ['LuaMouTiejiAttack2'] = '扰阵疲敌',
-    [':LuaMouTiejiAttack2'] = '若成功，你摸2张牌',
-    ['LuaMouTiejiDefense1'] = '拱位中军',
-    [':LuaMouTiejiDefense1'] = '阻止其获得你的一张牌',
-    ['LuaMouTiejiDefense2'] = '出阵应战',
-    [':LuaMouTiejiDefense2'] = '阻止其摸2张牌',
-    ['#LuaMouTiejiSuccess'] = '%from %arg <font color="green"><b>成功</b></font>',
-    ['#LuaMouTiejiFailure'] = '%from %arg <font color="red"><b>失败</b></font>',
     ['ExWenyang'] = '文鸯',
     ['&ExWenyang'] = '文鸯',
     ['#ExWenyang'] = '独骑破军',
