+++ lang/zh_CN/Package/StrategicAttackKnowledgePackage.lua
@@ -1,0 +1,45 @@
+-- translation for StrategicAttackKnowledgePackage
+
+return {
+    ['StrategicAttackKnowledgePackage'] = '谋攻篇-识包',
+    ['ExMouXuhuang'] = '谋徐晃',
+    ['&ExMouXuhuang'] = '谋徐晃',
+    ['#ExMouXuhuang'] = '径行截辎',
+    ['LuaMouDuanliang'] = '断粮',
+    [':LuaMouDuanliang'] = '出牌阶段限一次，你可以与一名其他角色进行一次谋弈：\
+    围城断粮，若其判定区中没有【兵粮寸断】，则你将牌堆顶的一张牌当无距离限制的【兵粮寸断】对其使用，否则你改为获得其一张牌；\
+    擂鼓进军，你视为对其使用一张【决斗】',
+    ['luamouduanliang'] = '断粮',
+    ['LuaMouDuanliangCard'] = '断粮',
+    ['LuaMouDuanliangMouyi'] = '断粮谋弈',
+    ['#LuaMouDuanliangSuccess'] = '%from %arg <font color="green"><b>成功</b></font>',
+    ['#LuaMouDuanliangFailure'] = '%from %arg <font color="red"><b>失败</b></font>',
+    ['LuaMouDuanliangAttack1'] = '围城断粮',
+    ['LuaMouDuanliangAttack2'] = '擂鼓进军',
+    ['LuaMouDuanliangDefense1'] = '全军突击',
+    ['LuaMouDuanliangDefense2'] = '闭门守城',
+    ['LuaMouShipo'] = '势迫',
+    [':LuaMouShipo'] = '结束阶段，你可以令一名体力值小于你的角色或所有判定区里有【兵粮寸断】的其他角色依次选择一项：1.弃置一张牌；2.令你摸一张牌',
+    ['LuaMouShipoChoice1'] = '令一名体力值小于你的角色进行选择',
+    ['LuaMouShipoChoice2'] = '令所有判定区里有【兵粮寸断】的其他角色依次选择',
+    ['LuaMouShipo-discard'] = '你需要弃置一张牌或者点击取消让 %src 摸一张牌',
+    ['LuaMouShipo-choose'] = '请选择一名其他角色，令其选择弃牌或是让你摸牌',
+    ['ExMouMachao'] = '谋马超',
+    ['&ExMouMachao'] = '谋马超',
+    ['#ExMouMachao'] = '阻戎负勇',
+    ['LuaMouTiejiMouyi'] = '铁骑谋弈',
+    ['LuaMouTieji'] = '铁骑',
+    [':LuaMouTieji'] = '当你使用【杀】指定一名角色为目标时，你可以令其所有非锁定技失效直到回合结束，且其不能使用【闪】响应此【杀】，使用然后你与该角色进行一次“谋弈”：\
+    1. 直取敌营，你获得其一张牌；\
+    2. 扰阵疲敌，你摸两张牌',
+    ['LuaMouTiejiAttack1'] = '直取敌营',
+    [':LuaMouTiejiAttack1'] = '若成功，你获得其一张牌',
+    ['LuaMouTiejiAttack2'] = '扰阵疲敌',
+    [':LuaMouTiejiAttack2'] = '若成功，你摸2张牌',
+    ['LuaMouTiejiDefense1'] = '拱位中军',
+    [':LuaMouTiejiDefense1'] = '阻止其获得你的一张牌',
+    ['LuaMouTiejiDefense2'] = '出阵应战',
+    [':LuaMouTiejiDefense2'] = '阻止其摸2张牌',
+    ['#LuaMouTiejiSuccess'] = '%from %arg <font color="green"><b>成功</b></font>',
+    ['#LuaMouTiejiFailure'] = '%from %arg <font color="red"><b>失败</b></font>'
+}
